move check_token_expiry() to services
